+++ R/model_parameters.R
@@ -5,6 +5,7 @@
 #'  \item{\link[=model_parameters.htest]{Correlations and t-tests}}
 #'  \item{\link[=model_parameters.aov]{ANOVAs}}
 #'  \item{\link[=model_parameters.lm]{Frequentist regressions} (\code{lm}, \code{glm}, ...)}
+#'  \item{\link[=model_parameters.zeroinfl]{Frequentist zero-inflated regressions} (\code{hurdle}, \code{zeroinfl})}
 #'  \item{\link[=model_parameters.merMod]{Frequentist mixed models} (\code{lme4}, \code{nlme}, \code{glmmTMB}, ...)}
 #'  \item{\link[=model_parameters.stanreg]{Bayesian models} (\code{rstanarm}, \code{brms})}
 #'  \item{\link[=model_parameters.principal]{PCA and FA} (\code{psych})}
@@ -48,4 +49,4 @@ parameters <- model_parameters
 # Summary
 # -------
 # In most cases model_parameters.polr() or model_parameters.lme() can be called
-# for new model objects.
+# for new model objects. A ".get_statistic()" function might be added.